Rename app to App
