+++ tests/Acceptance/App/Messenger/TestMessage.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace DR\SymfonyRequestId\Tests\Acceptance\App\Messenger;
+
+class TestMessage
+{
+}
