+++ tests/Acceptance/App/Monolog/MemoryHandler.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace DR\SymfonyRequestId\Tests\Acceptance;
+namespace DR\SymfonyRequestId\Tests\Acceptance\App\Monolog;
 
 use Countable;
 use DR\Utils\Assert;
