+++ tests/Acceptance/App/config/config.yml
@@ -6,14 +6,14 @@ framework:
   test: ~
   secret: "%secret%"
   router:
-    resource: "%kernel.project_dir%/app/config/routing.yml"
+    resource: "%kernel.project_dir%/config/routing.yml"
     strict_requirements: "%kernel.debug%"
     utf8: true
   messenger:
     transports:
       test_transport: 'sync://'
     routing:
-      DR\SymfonyRequestId\Tests\Acceptance\TestMessage: test_transport
+      DR\SymfonyRequestId\Tests\Acceptance\App\Messenger\TestMessage: test_transport
   http_method_override: false
   php_errors:
     log: true
@@ -28,7 +28,7 @@ symfony_request_id:
 twig:
   strict_variables: "%kernel.debug%"
   paths:
-    "%kernel.project_dir%/app/templates": requestid
+    "%kernel.project_dir%/templates": requestid
 
 services:
   log.formatter:
@@ -37,11 +37,11 @@ services:
       - "[%%level_name%% - %%extra.request_id%%] %%message%%"
   log.memory_handler:
     public: true
-    class: DR\SymfonyRequestId\Tests\Acceptance\MemoryHandler
+    class: DR\SymfonyRequestId\Tests\Acceptance\App\Monolog\MemoryHandler
     calls:
       - [ setFormatter, [ "@log.formatter" ] ]
   message.handler:
-    class: DR\SymfonyRequestId\Tests\Acceptance\TestMessageHandler
+    class: DR\SymfonyRequestId\Tests\Acceptance\App\Messenger\TestMessageHandler
     tags:
       - { name: messenger.message_handler, from_transport: test_transport, method: __invoke }
