--- tests/Acceptance/TestMessage.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace DR\SymfonyRequestId\Tests\Acceptance;
-
-class TestMessage
-{
-}
